pyproject.toml
[tool.poetry]
name = "bank-app-automation-testing"
version = "0.1.0"
description = ""
authors = ["Andrei Zrelov <qa.zrelov.aa@gmail.com>"]
readme = "README.md"
packages = [{include = "bank_app_automation_testing"}]

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.8.0"
pytest = "^7.2.1"
allure-pytest = "^2.12.0"
requests = "^2.28.2"
faker = "^16.6.1"
webdriver-manager = "^3.8.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
